[ConstraintSystem] NFC: Move `SolutionResult.h` to `include/swift/Sema`
