+++ lib/Sema/BuilderTransform.cpp
@@ -17,7 +17,6 @@
 
 #include "ConstraintSystem.h"
 #include "MiscDiagnostics.h"
-#include "SolutionResult.h"
 #include "TypeChecker.h"
 #include "TypeCheckAvailability.h"
 #include "swift/Sema/IDETypeChecking.h"
@@ -28,6 +27,7 @@
 #include "swift/AST/NameLookupRequests.h"
 #include "swift/AST/ParameterList.h"
 #include "swift/AST/TypeCheckRequests.h"
+#include "swift/Sema/SolutionResult.h"
 #include "llvm/ADT/DenseMap.h"
 #include "llvm/ADT/SmallVector.h"
 #include <iterator>
